// app/types/Customer.tsx
export type CustomerType = {
  id: number;
  customer_file_id: number;
  customer_contract_id: number | null;
  company_category_id: number;
  administrative_id: number;
  service_type_id: number;
  branch_id: number;
  company_id: number | null;
  tax_regime_id: number | null;
  general_sedes: number;
  name: string;
  phone: string;
  email: string;
  zip_code: number;
  city: string;
  state: string;
  address: string;
  status: string;
  map_location_url: string | null;
  tax_name: string | null;
  rfc: string | null;
  start_time: string | null;
  end_time: string | null;
  business_area_branch: string | null;
  meters: string | null;
  unit: string | null;
  url: string | null;
  portal_email: string | null;
  username: string | null;
  password: string | null;
  blueprints: number;
  registration_time: string;
  print_doc: number;
  validate_certificate: number;
  created_at: string;
  updated_at: string;
};